+++ Loader/PhpFileLoader.php
@@ -75,20 +75,8 @@ class PhpFileLoader extends FileLoader
         }, null, null);
 
         try {
-            try {
-                if (1 === $result = $load($path, $this->env)) {
-                    $result = null;
-                }
-            } catch (\Error $e) {
-                $load = \Closure::bind(function ($path, $env) use ($container, $loader, $resource, $type) {
-                    return include $path;
-                }, $this, ProtectedPhpFileLoader::class);
-
-                if (1 === $result = $load($path, $this->env)) {
-                    $result = null;
-                }
-
-                trigger_deprecation('symfony/dependency-injection', '8.1', 'Using `$this` or its internal scope in config files is deprecated, use the `$loader` variable instead in "%s" on line %d.', $e->getFile(), $e->getLine());
+            if (1 === $result = $load($path, $this->env)) {
+                $result = null;
             }
 
             if (\is_object($result) && \is_callable($result)) {